Moving civic tests in own directory
